Change project name from 'Proyecto' to 'proyecto' in frontend code
This commit updates the name of the project from 'Proyecto' to 'proyecto' in the frontend code. The changes include updating the name in various files such as angular.json, ionic.config.json, package-lock.json, and package.json. Additionally, the name of the folder containing the frontend code has also been changed from 'Proyecto' to 'proyecto'. The Dockerfile has been updated to reflect the new folder name.

The backend code has not been changed in this commit.